+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 ## Unreleased
+  - Fixed Describe's handling of TestName filter when multiple strings are passed to Invoke-Pester's -TestName parameter.
+  - Failing BeforeEach or AfterEach will fail the test [GH-326]
+  - Added BeOfType operator to the Should command. [GH-327]
+  - Fixed BeforeEach / etc parsing in PSv3+ so breakpoints and automatic variables ($PSCommandPath, etc) will work properly.  [GH-333]
+  - Fixed bug in 'Should Be' when comparing strings, and null or empty strings are piped in to the Should command.  [GH-333]
+  - Added some calls to Write-Progress in the It command.  [GH-322]
+  - Bug fix when mocking functions that are in the global scope; the original functions were being lost when the Describe block ended.  [GH-323]
+  - Improved failed assertion output from Assert-MockCalled; now behaves more like Should.  [GH-324]
+  - Added -ExclusiveFilter parameter to Assert-MockCalled.  Works like -ParameterFilter, except there also must not be any calls to the mocked command which do _not_ match the filter.
+  - Added the "bin" folder to the PATH environment variable when installing from Chocolatey.  Also removed the hard-coded -OutputXml and -Strict parameters from this file; only -EnableExit is always used from the bat file now.  [GH-281]
+  - PassThru object (when used in conjunction with -CodeCoverage) now includes information about Hit commands in addition to Missed commands.  [GH-341]
+
+## 3.3.8 (April 15, 2015)
+  - Further mocking fixes around the use of $ExecutionContext in client scope.  [GH-307]
+
+## 3.3.7 (April 15, 2015)
   - Added workaround for GetDynamicParameters() bug that was affecting mocks on the ActiveDirectory module in Windows 7. [GH-295]
+  - Revised Mocking code to avoid potential bugs when functions define parameters named $ExecutionContext or $MyInvocation. [GH-304]
+  - Mocked functions no longer call Get-MockDynamicParameters if the original function had no dynamicparam block. [GH-306]
 
 ## 3.3.6 (March 19, 2015)
   - Fix for mocking aliases for commands that are in scopes that Pester can't normally see. [GH-267]